move type files to root
